requirements.txt
streamlit
onnxruntime
numpy
torch
torchvision
requests
Pillow
streamlit-image-comparison
